Start Screen + Exposition Mechanics
Added a temp/barebones UI for a start screen. Paired with a camera animation that brings us in on the start.

The exposition's mechanics are done. There is a pitcher and three pebbles that the Crow has to pick up and bring to it. In this level, the crow can only pick up one pebble at a time (however in the next three Acts this is irrelevant since Crow is given pebbles by other critters).

Aside from the finished art assets, all that is necessary for this Exposition is the text from the script!